Limit the size of the haunted set. (#129)
The haunted set of the cache exists to provide temporal consistency of
keys that have been seen in the cache. This is because a read transaction
of an older transaction generation may attempt to include a value to the
cache that has been deleted or modified. To prevent this evicted items
are sent to the haunted list, which retains the key and the eviction
transaction id. This way any inclusion event that is older than the
caches haunted item transaction id will be ignored.

These haunted sets however can become very large with a high number of
keys and especially if the keys and values are nearly the same size.

As the value of the haunted items "diminishes" over time, and we already
have the ability to deny includes older than a minimum transaction id
(as part of cache clearing) we can re-use these mechanisms to trim the
haunted set.

This way the haunted set may only contain entries that are within
the look back limit, or the number of transaction in the past that
we allow to be used.

A lower look back limit means that the haunted sets are smaller, but
cache accuracy may be lower as we can no longer use older transactions
data. A higher look back limit means the haunted sets are larger but
cache accuracy can improve as we can include a range of transactions
inclusions and hit metrics.

The default value for this tries to balance this for highly parallel
and concurrent workloads. Workloads will long running transactions
will benefit from a larger look back limit, a work load with smaller
transaction lifes likely needs no changes.