Improve expected working set calculation
The previous working set size calculations were implemented when we
had unbounded tracking queues. This meant that a more complex
formula was used to account for queues, thread caches, and the shared
cache size.

However, we now use bounded queues which always provide a static limit
to memory consumption. This means that we can simplify the formula
for expected working set calculation.